Move project-wide from doubles to floats.
This not only fixes the performance regression from before, but improves render times.
For hdr.json we went from 45s to 36.4s average of 5 (Previous regression was 55s avg of 5)

+++ src/datatypes/camera.h
@@ -11,9 +11,9 @@
 struct renderer;
 
 struct camera {
-	double FOV;
-	double focalLength;
-	double aperture;
+	float FOV;
+	float focalLength;
+	float aperture;
 	
 	struct vector pos;
 	struct vector up;
